Sr. lognormal_dist now calculates all roots of the cubic equation
Changed subroutine lognormal_dist: Now it calculates all roots of the cubic equation (real and complex). The method is to construct an upper Hessenberg matrix whose eigenvalues are the desired roots, and then use the subroutines balanc and hqr (also added). The real and imaginary parts of the roots are returned in rtr(1:m) and rti(1:m), respectively. Then it chooses the real and positive one.