Topic: Added option for weighted mse loss training.
Description: Added new function weighted_mse_loss to models.py. It is a
mean squares loss function that weights the difference by the reciprocal
of the integration time. e.g. a one second trial with a difference of .2
in radius will have the same loss as a two second trial with a
difference of .4.  In general, the loss on a batch is:

loss = 1/n*((MLP_approximations - training_values)/integration_times)**2